MLOps_source_code/FlaskApplication/requirements.txt
# lightgbm==3.3.0
# numpy==1.18.5
# joblib==1.0.1
# seaborn==0.11.2
# matplotlib==3.4.2
# scikit-image==0.18.1
# scikit-learn==0.23.0
# shap==0.40.0
# xgboost==1.4.2
boto3==1.18.63
pandas==1.3.0
scikit_learn==0.24.2
Flask==2.0.1
pandas==1.3.0
gunicorn==20.1.0
python-dotenv==0.19.0